taskafarian/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==19.3.0             # via -r requirements.in, chalice
bcrypt==3.2.0             # via -r requirements.in
botocore==1.18.1          # via chalice
cffi==1.14.2              # via bcrypt
chalice==1.20.0           # via -r requirements.in
click==7.1.2              # via chalice
enum-compat==0.0.3        # via chalice
jmespath==0.10.0          # via botocore, chalice
marshmallow==3.8.0        # via -r requirements.in
mypy-extensions==0.4.3    # via chalice
psycopg2-binary==2.8.6    # via -r requirements.in
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via -r requirements.in
python-dateutil==2.8.1    # via botocore
pyyaml==5.3.1             # via chalice
six==1.15.0               # via bcrypt, chalice, python-dateutil
urllib3==1.25.10          # via botocore
wheel==0.35.1             # via chalice

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
